backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
alembic==1.14.0
psycopg[binary]==3.2.3
pydantic-settings==2.6.1
python-multipart==0.0.20
pypdf==5.1.0
openai==1.59.6
httpx==0.28.1
pytest==8.3.4
